changed linkvsp

+++ linkvsp.sh
@@ -4,13 +4,15 @@
 
 ctr=0
 for d in $(find /usr/local/Cellar -type d -name "vsp-*"); do
-  echo "Found $(basename $d)!"
+  #echo "Found $(basename $d)!"
   for f in $d/*/lib/*; do
-    fn=$(basename $f)
-    if [ ! -e /usr/local/lib/vapoursynth/$fn ]; then
-      echo "Missing symlink: $fn"
-      ln -s $f /usr/local/lib/vapoursynth/$fn
-      ctr=$((ctr+1))
+    if [[ ! -L $f && -f $f ]]; then
+      fn=$(basename $f)
+      if [ ! -e /usr/local/lib/vapoursynth/$fn ]; then
+        echo "Found missing symlink: $fn"
+        ln -s $f /usr/local/lib/vapoursynth/$fn
+        ctr=$((ctr+1))
+      fi
     fi
   done
 done
